Refactoring ContextProvider.js
