+++ src/components/layout/Navbar.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, {useState} from 'react';
 import {AiOutlineMenu} from 'react-icons/ai';
 import {BsFillPersonFill} from 'react-icons/bs';
 import {MdKeyboardArrowDown} from 'react-icons/md';
@@ -9,12 +9,13 @@ import {GrMoney} from "react-icons/gr";
 import {Link} from "react-router-dom";
 
 export default function Navbar() {
-    const {handleClick, isClicked, handleActiveMenu} = useStateContext();
+    const {activeMenu, setActiveMenu} = useStateContext();
     const {user} = useGetUserInfo();
+    const [isUserProfileOpen, setUserProfileOpen] = useState(false);
 
     return (
         <div className="flex w-full justify-between p-1 relative z-40 bg-white">
-            <button type="button" onClick={handleActiveMenu}
+            <button type="button" onClick={() => setActiveMenu(!activeMenu)}
                     className="main-color relative text-xl rounded-full p-3 hover:bg-light-gray"
             ><AiOutlineMenu/></button>
             <div className="flex">
@@ -25,13 +26,13 @@ export default function Navbar() {
                         <span className="text-gray-700 font-bold ml-1 text-14">₩{user.point.toLocaleString('ko-KR')}</span>
                     </p>
                 </Link>
-                <div onClick={() => handleClick('userProfile')}
+                <div onClick={() => setUserProfileOpen((prevState) => !prevState)}
                      className="flex items-center gap-2 cursor-pointer px-2 hover:bg-light-gray rounded-lg">
                     <BsFillPersonFill className="w-5 h-5"/>
                     <span className="text-gray-700 font-bold text-14">{user.username}</span>
                     <MdKeyboardArrowDown className="text-gray-400 text-14"/>
                 </div>
-                {isClicked.userProfile && (<UserProfile/>)}
+                {isUserProfileOpen && (<UserProfile/>)}
             </div>
         </div>
     );
