+++ src/components/layout/Sidebar.tsx
@@ -6,7 +6,10 @@ import {GiChameleonGlyph} from 'react-icons/gi';
 import {MdOutlineCancel} from 'react-icons/md';
 
 export default function Sidebar() {
-    const {activeMenu, setActiveMenu, handleCloseSideBar} = useStateContext();
+    const {activeMenu, setActiveMenu} = useStateContext();
+    const handleCloseSideBar = () => {
+        if (activeMenu !== undefined) setActiveMenu(false);
+    };
 
     const activeLink = 'flex items-center gap-5 pl-4 pt-3 pb-2.5 rounded-lg text-white text-md m-2';
     const normalLink = 'flex items-center gap-5 pl-4 pt-3 pb-2.5 rounded-lg text-md text-gray-700 hover:bg-light-gray m-2';
